source-code/backend/requirements.txt
Flask==3.1.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.1.2
gunicorn==22.0.0